Fix some unintentional exports of private names.
* move version.py to jax/_src/version.py and add a shim that reexports only specific names.
* explicitly delete a few deprecated names.
