+++ jax/_src/BUILD
@@ -35,6 +35,10 @@ exports_files(
     visibility = jax_export_file_visibility,
 )
 
+exports_files(
+    ["version.py"],
+)
+
 pytype_strict_library(
     name = "init",
     srcs = [
@@ -720,6 +724,11 @@ pytype_strict_library(
     name = "deprecations",
     srcs = ["deprecations.py"],
     type_checking = "pyrefly",
+)
+
+pytype_strict_library(
+    name = "version",
+    srcs = ["version.py"],
 )
 
 pytype_strict_library(
