+++ jax/_src/lib/BUILD
@@ -35,7 +35,7 @@ py_library_providing_imports_info(
     type_checking = "pyrefly",
     visibility = ["//jax:internal"] + jax_visibility("lib"),
     deps = [
-        "//jax:version",
+        "//jax/_src:version",
         "//jax/_src:lazy_loader",
     ] + if_building_jaxlib([
         "//jaxlib",
