+++ jax/scipy/special.py
@@ -95,4 +95,6 @@ else:
   from jax._src.deprecations import deprecation_getattr as _deprecation_getattr
   __getattr__ = _deprecation_getattr(__name__, _deprecations)
   del _deprecation_getattr
+  del _deprecated_lpmn
+  del _deprecated_lpmn_values
 del _typing
