+++ tests/mosaic/gpu_layout_inference_test.py
@@ -19,6 +19,7 @@ import dataclasses
 
 from absl.testing import parameterized
 import jax
+import jax._src.version
 from jax import numpy as jnp
 from jax._src import config
 from jax._src import dtypes
@@ -110,7 +111,7 @@ class LayoutInferenceTest(parameterized.TestCase):
     ]
 
   def setUp(self):
-    if jax.version._version != jax.lib.__version__:
+    if jax._src.version._version != jax.lib.__version__:
       raise self.skipTest("Test requires matching jax and jaxlib versions")
     super().setUp()
     if jtu.test_device_matches(["rocm"]):
